# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Access live interpreter platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  message: string;
}

export type Role = typeof Role[keyof typeof Role];


export const Role = {
  admin: 'admin',
  user: 'user',
  interpreter: 'interpreter',
} as const;

export type InterpreterStatus = typeof InterpreterStatus[keyof typeof InterpreterStatus];


export const InterpreterStatus = {
  available: 'available',
  busy: 'busy',
  offline: 'offline',
} as const;

export type SessionStatus = typeof SessionStatus[keyof typeof SessionStatus];


export const SessionStatus = {
  pending: 'pending',
  active: 'active',
  ended: 'ended',
  declined: 'declined',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: Role;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Language {
  id: number;
  code: string;
  name: string;
  nameAr: string;
  flagEmoji: string;
}

export interface LanguageInput {
  code: string;
  name: string;
  nameAr: string;
  flagEmoji: string;
}

export interface InterpreterProfile {
  id: number;
  name: string;
  email: string;
  status: InterpreterStatus;
  languages: Language[];
}

export interface Interpreter {
  id: number;
  name: string;
  email: string;
  status: InterpreterStatus;
  languages: Language[];
}

export interface Session {
  id: number;
  userId: number;
  userName: string;
  /** @nullable */
  interpreterId?: number | null;
  /** @nullable */
  interpreterName?: string | null;
  languageId: number;
  language: Language;
  status: SessionStatus;
  createdAt: string;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  endedAt?: string | null;
}

export interface SessionRequestBody {
  languageId: number;
}

export interface InterpreterStatusBody {
  status: InterpreterStatus;
}

export interface UserInput {
  name: string;
  email: string;
  password: string;
}

export interface InterpreterInput {
  name: string;
  email: string;
  password: string;
  languageIds: number[];
}

export interface SetLanguagesBody {
  languageIds: number[];
}

export interface AdminStats {
  totalUsers: number;
  totalInterpreters: number;
  availableInterpreters: number;
  activeSessions: number;
  completedSessions: number;
}

export type ListLanguagesParams = {
/**
 * When true, only return languages that have at least one interpreter assigned to them (coverage), regardless of current availability. Used by the user-facing language picker.
 */
assigned?: boolean;
};

export type ListAdminSessionsParams = {
status?: SessionStatus;
};